[Aptos Data Client] Split poller into separate file.
